+++ js/homeview.js
@@ -111,7 +111,8 @@ enyo.kind({
 	// Init activities service response, redraw screen
 	queryActivitiesResponse: function(inSender, inResponse) {
 		// No activities at start
-		if (preferences.getActivities() == null) {
+		var currentList = preferences.getActivities();
+		if (currentList == null || currentList.length == 0) {
 			// Just copy the activities from the service
 			preferences.setActivities(inResponse.data);
 			preferences.save();
@@ -697,7 +698,6 @@ enyo.kind({
 			this.otherview = this.$.otherview.createComponent({kind: "Sugar.DialogWarningMessage"}, {owner:this});
 			this.otherview.show();
 		} else {
-			preferences.addUserInHistory();
 			util.cleanDatastore(null, function() {
 				util.restartApp();
 			});