+++ CHANGELOG.md
@@ -9,11 +9,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fraction activity #205
 - Implode activity #710
 - Planets activity
+- Curriculum activity
 - Simon mode game in TamTam activity #686
 - Vue.js activity template
 - Activity Development Tutorial in Vue.js
 - Duplicate action on Journal entries
 - ES6 compatibility for activities (except modules)
+- Support autologoff mode for Electron and Sugarizer OS
+- Support for native Android app in activities.json
 
 ### Changed
 - Improve Presence palette UI, now handle list of users and disconnection
@@ -54,6 +57,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Last One Loses activity is unplayable after sharing without second player #815
 - Spinner on first screen is at the bottom of the screen
 - Countries and view palettes pop at the wrong place in Constellation activity
+- Timer in chess activity stops on inactive tabs #828
 
 
 ## [1.3.0] - 2020-03-28